Setup new Tool to take a Model 3D and Convert it for our 3D Library to load
