+++ AmberDev.code-workspace
@@ -7,7 +7,7 @@
 			"path": "RPG_DEV"
 		},
 		{
-			"path": "Tools"
+			"path": "Model3D_Converter"
 		}
 	],
 	"settings": {
