+++ Model3D_Converter/.vscode/c_cpp_properties.json
@@ -4,6 +4,7 @@
             "name": "Linux",
             "includePath": [
                 "${workspaceFolder}/**",
+                "${workspaceFolder}/../Amber3D_SL/**",
                 "/usr/include/qt/**"
             ],
             "defines": [],
