+++ Model3D_Converter/.vscode/tasks.json
@@ -4,13 +4,16 @@
     "version": "2.0.0",
     "tasks": [
         {
-            "label": "Linux Build",
+            "label": "Model3D_Converter Build",
             "type": "shell",
             "command": "./build.sh",
             "group": {
                 "kind": "build",
                 "isDefault": true
-            }
+            },
+            "problemMatcher": [
+                "$gcc"
+            ]
         }
     ]
 }
