+++ Model3D_Converter/ConverterWindow.cpp
@@ -1,0 +1,15 @@
+#include "ConverterWindow.h"
+#include "ui_ConverterWindow.h"
+
+ConverterWindow::ConverterWindow(QWidget *parent)
+    : QMainWindow(parent)
+    , ui(new Ui::ConverterWindow)
+{
+    ui->setupUi(this);
+}
+
+ConverterWindow::~ConverterWindow()
+{
+    delete ui;
+}
+
