+++ Model3D_Converter/build.sh
@@ -1,3 +1,4 @@
 cd build
+#make clean
 qmake ../
 make
