+++ Model3D_Converter/main.cpp
@@ -1,0 +1,11 @@
+#include "ConverterWindow.h"
+
+#include <QApplication>
+
+int main(int argc, char *argv[])
+{
+    QApplication a(argc, argv);
+    ConverterWindow w;
+    w.show();
+    return a.exec();
+}
