--- Tools/User_Data/DataWindow.ui
@@ -1,101 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>DataWindow</class>
- <widget class="QMainWindow" name="DataWindow">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>1000</width>
-    <height>1000</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>DataWindow</string>
-  </property>
-  <widget class="QWidget" name="centralwidget">
-   <widget class="AmberData::UserData" name="Speadsheet">
-    <property name="geometry">
-     <rect>
-      <x>10</x>
-      <y>10</y>
-      <width>1100</width>
-      <height>861</height>
-     </rect>
-    </property>
-   </widget>
-   <widget class="QPushButton" name="SaveButton">
-    <property name="geometry">
-     <rect>
-      <x>90</x>
-      <y>940</y>
-      <width>80</width>
-      <height>26</height>
-     </rect>
-    </property>
-    <property name="text">
-     <string>Save</string>
-    </property>
-   </widget>
-   <widget class="QPushButton" name="LoadButton">
-    <property name="geometry">
-     <rect>
-      <x>90</x>
-      <y>900</y>
-      <width>80</width>
-      <height>26</height>
-     </rect>
-    </property>
-    <property name="text">
-     <string>Load</string>
-    </property>
-   </widget>
-   <widget class="QPushButton" name="RefreshButton">
-    <property name="geometry">
-     <rect>
-      <x>280</x>
-      <y>900</y>
-      <width>80</width>
-      <height>26</height>
-     </rect>
-    </property>
-    <property name="text">
-     <string>Refresh</string>
-    </property>
-   </widget>
-   <widget class="QPushButton" name="ExitButton">
-    <property name="geometry">
-     <rect>
-      <x>850</x>
-      <y>940</y>
-      <width>80</width>
-      <height>26</height>
-     </rect>
-    </property>
-    <property name="text">
-     <string>Exit</string>
-    </property>
-   </widget>
-   <widget class="QWidget" name="formLayoutWidget">
-    <property name="geometry">
-     <rect>
-      <x>610</x>
-      <y>450</y>
-      <width>160</width>
-      <height>80</height>
-     </rect>
-    </property>
-    <layout class="QFormLayout" name="formLayout"/>
-   </widget>
-  </widget>
- </widget>
- <customwidgets>
-  <customwidget>
-   <class>AmberData::UserData</class>
-   <extends>QTableWidget</extends>
-   <header>Source/CurrentData/Data.h</header>
-  </customwidget>
- </customwidgets>
- <resources/>
- <connections/>
-</ui>
