--- Tools/User_Data/Source/CurrentData/BaseData.dat
@@ -1,50 +1,0 @@
-User Name									
-Avatar Name									
-Avatar Tile									
-Homepage									
-Strength	9	Fatigue	9	Thrust	1d-2	Swing	1d-1		
-Intelligence	9	Psi Points	9						
-Dexterity	9								
-Health	9	Hits Taken	0	Base Move	4.5				
-Encumberance	0	Amour Defence	0	Shield Defence	0	Passive Defence	0		
-Doge	4.5	Parry	0	Block	0				
-Abilities	0	Skills	0	Combat Skills	0	Psi Skills	0		
-Equipment	0	Cash	0	Bank	0	Storage	0		
-Properties	0								
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-									
-
